fix 64bit warnings (use: GYP_DEFINES=skia_arch_width=64 make ... )
git-svn-id: http://skia.googlecode.com/svn/trunk@4464 2bbb7eff-a529-9590-31e7-b0007b416f81

+++ include/core/SkChecksum.h
@@ -110,8 +110,8 @@ public:
          *  if we're on a 32bit or 64bit arch, since we use sizeof(uintptr_t)
          *  to compute how much to shift-down the size.
          */
-        int n4 = size / (sizeof(uintptr_t) << 2);
-        for (int i = 0; i < n4; ++i) {
+        size_t n4 = size / (sizeof(uintptr_t) << 2);
+        for (size_t i = 0; i < n4; ++i) {
             result = Mash(result, *ptr++);
             result = Mash(result, *ptr++);
             result = Mash(result, *ptr++);
